requirements.txt
python-telegram-bot==13.15
Flask
requests
python-dotenv
APScheduler
scikit-learn
joblib
numpy
scipy
six==1.16.0
urllib3==1.26.18
